Disable double promotion

+++ STM32/Src/BiquadDesigner/biquad.h
@@ -19,12 +19,12 @@ typedef struct iir_filter {
 
 iir_filter_t *biquad_create(int sections);
 void biquad_delete(iir_filter_t *filter);
-double biquad_update(iir_filter_t *filter, double x);
-void iir_freq_resp(iir_filter_t *filter, double *h, double fs, double f);
+double biquad_update(iir_filter_t *filter, float x);
+void iir_freq_resp(iir_filter_t *filter, float *h, float fs, float f);
 void biquad_zero(struct iir_filter *filter);
-void biquad_init_lowpass(iir_filter_t *filter, double fs, double f);
-void biquad_init_highpass(iir_filter_t *filter, double fs, double f);
-void biquad_init_bandpass(iir_filter_t *filter, double fs, double f1, double f2);
-void biquad_init_bandstop(iir_filter_t *filter, double fs, double f1, double f2);
+void biquad_init_lowpass(iir_filter_t *filter, float fs, float f);
+void biquad_init_highpass(iir_filter_t *filter, float fs, float f);
+void biquad_init_bandpass(iir_filter_t *filter, float fs, float f1, float f2);
+void biquad_init_bandstop(iir_filter_t *filter, float fs, float f1, float f2);
 
 #endif
